requirements.txt
tk
google-generativeai
requests
beautifulsoup4
pywin32